refactor: split UpdateSecretCommand god file (854->659 lines)
Extract two cohesive units from UpdateSecretCommand.cs and dedupe the
non-interactive success tail. No behavior change; 194/194 tests green.

- ImpersonationRunner.cs: RunImpersonated + LogonUser P/Invoke + logon
  constants, isolating the native/extern surface from CLI dispatch.
- InteractiveConfigPrompts.cs: the Prompt* helpers + ReadMaskedInput.
  PromptSecret now takes a pre-resolved secret so the prompts unit has no
  back-dependency on the command (command resolves --secret/-file/-env).
- ReportConfigWritten: shared tail for RunSilent/RunSilentImpersonated.

The JenkinsAgentWorker process-launch methods were deliberately left in
place: they read/write ~10 instance fields and sit on hard security
invariants (secret-file cleanup path, log redaction), so extraction would
trade a documented-invariant risk for a line-count win.
