+++ CLAUDE.md
@@ -39,7 +39,7 @@ The `.slnx` builds the app + tests only; the installer has `Build=false` and mus
 
 - **Host / entry:** `Program.cs` (top-level statements: CLI dispatch → Serilog → host/DI → OTel → pre-flight validate → run), `JenkinsAgentWorker.cs` (`BackgroundService` — lifecycle, supervision loop, watchdog, metrics)
 - **Config POCOs:** `ServiceSettings.cs` (nested: `Connection`/`Secret`/`Agent`/`Hardening`/`Logging`/`Recovery`), `TelemetrySettings.cs`, `ConfigKeys.cs` (centralized section/key name constants — **use these, don't hardcode config strings**), `ConnectionMethod.cs`, `SecretMode.cs`, `DpapiScope.cs`, `ServiceSettingsNormalizer.cs` (reconciles `appsettings.json` to the current schema on MSI upgrade: adds missing keys at POCO defaults, prunes keys/sections the schema no longer defines, preserves in-schema values)
-- **Secrets:** `ISecretResolver.cs`/`SecretResolver.cs`, `SecretWriter.cs`, `UpdateSecretCommand.cs` (the `update-secret` CLI), `TpmSecretProtector.cs`, `AgentSecretFile.cs` (`-secret @<file>` off-argv), `CertificateThumbprintValidator.cs`
+- **Secrets:** `ISecretResolver.cs`/`SecretResolver.cs`, `SecretWriter.cs`, `UpdateSecretCommand.cs` (the `update-secret` CLI — dispatch, arg parsing, silent/upgrade flows), `InteractiveConfigPrompts.cs` (console prompt primitives for the interactive flow, incl. masked input), `ImpersonationRunner.cs` (`LogonUser` + `RunImpersonated` for Credential-Manager/User-DPAPI writes; isolates the native `extern` surface), `TpmSecretProtector.cs`, `AgentSecretFile.cs` (`-secret @<file>` off-argv), `CertificateThumbprintValidator.cs`
 - **Agent process:** `AgentProcessLauncher.cs` (the `IAgentProcessLauncher`/`IAgentProcess` seam — production wraps `System.Diagnostics.Process`; fakes drive the watchdog in tests), `AgentArgumentParser.cs`, `AgentTransport.cs` (pure transport-selection: `InitialMethod`/`Toggle`/`ShouldFallback`), `JavaPathResolver.cs`, `ServiceSettingsValidator.cs`
 - **Networking:** `IJarDownloader.cs`/`HttpJarDownloader.cs` (ETag conditional GET; temp-file + atomic move so a truncated download can't sit behind a valid ETag; records the jar's SHA-256 on download and re-verifies the cached jar on each 304 — a mismatch forces an unconditional re-download. TOFU/local-integrity, not upstream authenticity — that's `Connection:ControllerCertThumbprint`), `IConnectivityChecker.cs`/`TcpConnectivityChecker.cs`
 - **Hardening:** `ProcessMitigations.cs`, `EnvironmentSanitizer.cs`, `ConfigAclHardener.cs`, `EventLogSourceInstaller.cs`, `DataPaths.cs` (binary/data split)
